move mocker file to sep file
